+++ server/src/main/java/org/elasticsearch/common/scheduler/SchedulerEngine.java
@@ -1,11 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.core.scheduler;
+package org.elasticsearch.common.scheduler;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -30,6 +31,12 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
+/**
+ * Thread-safe scheduling implementation that'll cancel an already scheduled job
+ * before rescheduling.
+ * The scheduler engine offers support for registering listeners that'll get notified
+ * when a job triggers (according to the job's configured schedule)
+ */
 public class SchedulerEngine {
 
     public static class Job {
