+++ server/src/main/java/org/elasticsearch/common/scheduler/TimeValueSchedule.java
@@ -1,17 +1,22 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.ilm;
+package org.elasticsearch.common.scheduler;
 
+import org.elasticsearch.common.scheduler.SchedulerEngine.Schedule;
 import org.elasticsearch.core.TimeValue;
-import org.elasticsearch.xpack.core.scheduler.SchedulerEngine.Schedule;
 
 import java.util.Objects;
 
+/**
+ * {@link Schedule} implementation wrapping a {@link TimeValue} interval that'll compute the
+ * next scheduled execution time according to the configured interval.
+ */
 public class TimeValueSchedule implements Schedule {
 
     private TimeValue interval;
