+++ x-pack/plugin/core/src/main/java/org/elasticsearch/license/ExpirationCallback.java
@@ -7,6 +7,7 @@
 package org.elasticsearch.license;
 
 import org.elasticsearch.common.logging.LoggerMessageFormat;
+import org.elasticsearch.common.scheduler.SchedulerEngine;
 import org.elasticsearch.core.TimeValue;
 
 import java.util.UUID;
@@ -138,7 +139,7 @@ abstract class ExpirationCallback {
     }
 
     /**
-     * {@link org.elasticsearch.xpack.core.scheduler.SchedulerEngine.Schedule#nextScheduledTimeAfter(long, long)}
+     * {@link SchedulerEngine.Schedule#nextScheduledTimeAfter(long, long)}
      * with respect to license expiry date
      */
     public final long nextScheduledTimeForExpiry(long expiryDate, long startTime, long time) {
