+++ x-pack/plugin/core/src/test/java/org/elasticsearch/license/LicenseScheduleTests.java
@@ -6,9 +6,9 @@
  */
 package org.elasticsearch.license;
 
+import org.elasticsearch.common.scheduler.SchedulerEngine;
 import org.elasticsearch.core.TimeValue;
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.core.scheduler.SchedulerEngine;
 import org.junit.Before;
 
 import java.time.Instant;
